requirements.txt
discord.py==1.7.3
boto3==1.21.46
